Abstract block cache
