+++ src/main/java/org/opensearch/index/store/block_cache/BlockCache.java
@@ -5,18 +5,19 @@
 package org.opensearch.index.store.block_cache;
 
 import java.io.IOException;
+import java.util.Optional;
 
-public interface BlockCache {
+public interface BlockCache<T> {
 
     /**
      * Returns the block if cached, or null if absent.
      */
-    BlockCacheValue get(BlockCacheKey key);
+    BlockCacheValue<T> get(BlockCacheKey key);
 
     /**
      * Returns the block, loading it via `BlockLoader` if absent.
      */
-    BlockCacheValue getOrLoad(BlockCacheKey key, int size) throws IOException;
+    Optional<BlockCacheValue<T>> getOrLoad(BlockCacheKey key, int size) throws IOException;
 
     /**
      * Asynchronously load the block into the cache if not present.
@@ -26,7 +27,7 @@ public interface BlockCache {
     /**
      * Put a block into the cache.
      */
-    void put(BlockCacheKey key, BlockCacheValue value);
+    void put(BlockCacheKey key, BlockCacheValue<T> value);
 
     /**
      * Evict a block from the cache.
